# v3: Improve performance of Adaptor Middleware