--- a/middleware/adaptor/adaptor.go
+++ b/middleware/adaptor/adaptor.go
@@ -5,6 +5,7 @@ import (
 	"net"
 	"net/http"
 	"reflect"
+	"sync"
 	"unsafe"
 
 	"github.com/gofiber/fiber/v3"
@@ -12,6 +13,18 @@ import (
 	"github.com/valyala/fasthttp"
 	"github.com/valyala/fasthttp/fasthttpadaptor"
 )
+
+var reqPool = sync.Pool{
+	New: func() any {
+		return new(fasthttp.Request)
+	},
+}
+
+var ctxPool = sync.Pool{
+	New: func() any {
+		return new(fasthttp.RequestCtx)
+	},
+}
 
 // HTTPHandlerFunc wraps net/http handler func to fiber handler
 func HTTPHandlerFunc(h http.HandlerFunc) fiber.Handler {
@@ -82,12 +95,13 @@ func HTTPMiddleware(mw func(http.Handler) http.Handler) fiber.Handler {
 	return func(c fiber.Ctx) error {
 		var next bool
 		nextHandler := http.HandlerFunc(func(_ http.ResponseWriter, r *http.Request) {
-			next = true
 			// Convert again in case request may modify by middleware
+			next = true
 			c.Request().Header.SetMethod(r.Method)
 			c.Request().SetRequestURI(r.RequestURI)
 			c.Request().SetHost(r.Host)
 			c.Request().Header.SetHost(r.Host)
+
 			for key, val := range r.Header {
 				for _, v := range val {
 					c.Request().Header.Set(key, v)
@@ -124,9 +138,10 @@ func FiberApp(app *fiber.App) http.HandlerFunc {
 
 func handlerFunc(app *fiber.App, h ...fiber.Handler) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		// New fasthttp request
-		req := fasthttp.AcquireRequest()
-		defer fasthttp.ReleaseRequest(req)
+		req := reqPool.Get().(*fasthttp.Request) //nolint:forcetypeassert,errcheck // overlinting
+		req.Reset()
+		defer reqPool.Put(req)
+
 		// Convert net/http -> fasthttp request
 		if r.Body != nil {
 			n, err := io.Copy(req.BodyWriter(), r.Body)
@@ -141,26 +156,34 @@ func handlerFunc(app *fiber.App, h ...fiber.Handler) http.HandlerFunc {
 		req.SetRequestURI(r.RequestURI)
 		req.SetHost(r.Host)
 		req.Header.SetHost(r.Host)
+
 		for key, val := range r.Header {
 			for _, v := range val {
 				req.Header.Set(key, v)
 			}
 		}
+
 		if _, _, err := net.SplitHostPort(r.RemoteAddr); err != nil && err.(*net.AddrError).Err == "missing port in address" { //nolint:errorlint, forcetypeassert // overlinting
 			r.RemoteAddr = net.JoinHostPort(r.RemoteAddr, "80")
 		}
+
 		remoteAddr, err := net.ResolveTCPAddr("tcp", r.RemoteAddr)
 		if err != nil {
 			http.Error(w, utils.StatusMessage(fiber.StatusInternalServerError), fiber.StatusInternalServerError)
 			return
 		}
 
-		// New fasthttp Ctx
-		var fctx fasthttp.RequestCtx
+		// New fasthttp Ctx from pool
+		fctx := ctxPool.Get().(*fasthttp.RequestCtx) //nolint:forcetypeassert,errcheck // overlinting
+		fctx.Response.Reset()
+		defer ctxPool.Put(fctx)
 		fctx.Init(req, remoteAddr, nil)
+
 		if len(h) > 0 {
 			// New fiber Ctx
-			ctx := app.AcquireCtx(&fctx)
+			ctx := app.AcquireCtx(fctx)
+			defer app.ReleaseCtx(ctx)
+
 			// Execute fiber Ctx
 			err := h[0](ctx)
 			if err != nil {
@@ -168,10 +191,10 @@ func handlerFunc(app *fiber.App, h ...fiber.Handler) http.HandlerFunc {
 			}
 		} else {
 			// Execute fasthttp Ctx though app.Handler
-			app.Handler()(&fctx)
+			app.Handler()(fctx)
 		}
 
-		// Convert fasthttp Ctx > net/http
+		// Convert fasthttp Ctx -> net/http
 		fctx.Response.Header.VisitAll(func(k, v []byte) {
 			w.Header().Add(string(k), string(v))
 		})