# Description

Improve performance of the Adaptor middleware by using `sync.Pool` to re-use `fasthttp.RequestCtx` objects.

Overall Percent Change
- Execution Time: Reduced by ~47.57%
- Memory Usage: Reduced by ~89.97%
- Allocations: Reduced by ~68.75%

## Summary of Results
### Run 1 (Current main branch):

- 100KB: 1056 ns/op, 2644 B/op, 16 allocs/op
- 500KB: 1061 ns/op, 2644 B/op, 16 allocs/op
- 1MB: 1080 ns/op, 2646 B/op, 16 allocs/op
- 5MB: 1093 ns/op, 2654 B/op, 16 allocs/op
- 10MB: 1044 ns/op, 2665 B/op, 16 allocs/op
- 25MB: 1069 ns/op, 2706 B/op, 16 allocs/op
- 50MB: 1137 ns/op, 2734 B/op, 16 allocs/op

### Run 2 (This pull request)

- 100KB: 588.6 ns/op, 254 B/op, 5 allocs/op
- 500KB: 562.9 ns/op, 248 B/op, 5 allocs/op
- 1MB: 629.7 ns/op, 267 B/op, 5 allocs/op
- 5MB: 540.3 ns/op, 254 B/op, 5 allocs/op
- 10MB: 533.1 ns/op, 258 B/op, 5 allocs/op
- 25MB: 540.7 ns/op, 289 B/op, 5 allocs/op
- 50MB: 554.6 ns/op, 298 B/op, 5 allocs/op

## Type of change

- [x] Performance improvement (non-breaking change which improves efficiency)